requirements.txt
Flask 
pydantic
langchain
openai
langchain-openai
python-dotenv